requirements.txt
tornado==3.0.1
ujson==1.30
python-memcached==1.53
ipaddr==2.1.10
sh==1.08
futures==2.1.3
Pillow==2.2.2
pymongo==2.6.3
celery==3.0.22